runner: add `cases` functionality
This will allow us to specify and run various test cases by the runner
in order to record metrics about each specified test case and how it has
changed over the course of two revisions.

This commit adds the data structure `TestCase` which describes a single test
cases and all of the possible configuration and metadata options that a test
case can contain. Further, a `TestCasFile` model is also added which functions
as the "parsing" mechanism for the test cases files.