add mechanism and data structures for interpreting results
This commit adds all the necessary data structures to digest any
results that come from the test cases. The main structure is `ResultEntry`
which stores information about the `Metrics` of both the `original` and
`result` cases. As a utility, it also builds a `comparison` view, in order
for output methods to elegantly display differences in the data sets.